+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - Conversion tool: added new parser for DefiTaxes.us.
 - Conversion tool: allow parsers to specify a different worksheet
 - Conversion tool: added parser for BloFin exchange.
+- Coinbase parser: added "Subscription" transaction type.
+- Conversion tool: added parser for Coinbase Prime exchange.
 ### Changed
 - Config: fiat_income to True.
 - Price tool: CoinDesk API deprecated.